+++ lib/camera.h
@@ -10,14 +10,14 @@
 
 typedef struct Camera {
   size_t hsize, vsize; // in pixels
-  float fov;           // an angle
-  float pixel_size;    // of canvas, at 1 unit distance
-  float half_width, half_height;
+  double fov;          // an angle
+  double pixel_size;   // of canvas, at 1 unit distance
+  double half_width, half_height;
   Mat4 transform;
 } Camera;
 
 Mat4 view_transform(Point, Point, Vector);
-Camera camera(size_t, size_t, float);
+Camera camera(size_t, size_t, double);
 Ray ray_for_pixel(Camera, size_t, size_t);
 Canvas render(Camera, World);
